Add additional maintainers to avoid single point of failure
This feedstock only has a single maintainer. To avoid the case where this maintainer is unavailable to attend to duties related to this repo because of other engangements, I suggest adding two additional maintainers. The first is @mreineck, who is the lead developer of ducc0. The second is myself (@markwieczorek), who makes use of this project for the pyshtools project. @ickc, if you have any questions, please feel free to ask.